requirements.txt
django
redis
channels_redis

